Clear compiler warnings that are not Gson null guards
- DownloadProgressBar moves off the deprecated Float LinearProgressIndicator.
  A Robolectric test renders it next to the old call and compares pixels; it
  catches a 42 -> 50 progress change, and the old and new overloads draw
  identically on material3 1.3.1.
- Unit tests can now run Compose UI on the JVM (ui-test-junit4, merged
  manifest via isIncludeAndroidResources).
- UpdateService: OkHttp 5 made Response.body non-null, so the null-body branch
  was unreachable.
- Tests: redundant conversions and safe calls, deprecated Locale constructor.

Left alone: the safe calls on DTO fields. Gson can put null into a non-null
Kotlin property, so those guards are doing real work until the DTOs declare
the nullability the server actually has.

+++ app/src/test/java/com/paulohenriquesg/fahrenheit/api/PlaybackPayloadShapeTest.kt
@@ -62,7 +62,7 @@ class PlaybackPayloadShapeTest {
             PlayLibraryItemResponse::class.java
         )
 
-        assertEquals(31124.073333, response.audioTracks[0].startOffset.toDouble(), 1e-6)
+        assertEquals(31124.073333, response.audioTracks[0].startOffset, 1e-6)
     }
 
     @Test
@@ -75,7 +75,7 @@ class PlaybackPayloadShapeTest {
             PlayLibraryItemResponse::class.java
         )
 
-        assertEquals(3_000_000_000L, response.audioTracks[0].metadata.size.toLong())
+        assertEquals(3_000_000_000L, response.audioTracks[0].metadata.size)
     }
 
     @Test
@@ -112,8 +112,8 @@ class PlaybackPayloadShapeTest {
             PlayLibraryItemResponse::class.java
         )
 
-        assertEquals(19141.935520695628, response.currentTime.toDouble(), 1e-9)
-        assertEquals(538.6500963920058, response.startTime.toDouble(), 1e-9)
-        assertEquals(29.154561968609244, response.timeListening.toDouble(), 1e-9)
+        assertEquals(19141.935520695628, response.currentTime, 1e-9)
+        assertEquals(538.6500963920058, response.startTime, 1e-9)
+        assertEquals(29.154561968609244, response.timeListening, 1e-9)
     }
 }
